config: extract security-log compilation to compiler_security_log.go
Part B of #4121 (pure code motion). Moves the security-log cluster
verbatim out of compiler_security.go:

  - compileLog (streams + #2008 H7 log profiles)
  - validateSecurityLogStreamPortsAST (#3349 dual-location port gate)
  - validateSecurityLogStreamTLSProfileAST (#3350 unapplied-profile
    reject gate)

Function bodies are byte-identical to origin/master. The new file
imports fmt and strconv (the union the moved declarations use). All
declarations stay in package config so the commit-path callers of the
two AST validators resolve across the same package unchanged.

Validation: gofmt -l clean, go build ./pkg/config/ green.
